move code for booleanfield to separate file
