+++ tests/test_orm_core.py
@@ -32,7 +32,7 @@ from clorm import ( BaseField, Raw, RawField, IntegerField, StringField,
                     HeadList, HeadListReversed, TailList, TailListReversed )
 
 # Implementation imports
-from clorm.orm.core import ( BooleanField, StrictBooleanField, dealiased_path, field, get_field_definition,
+from clorm.orm.core import ( dealiased_path, field, get_field_definition,
                              PredicatePath, QCondition, trueall, notcontains )
 
 import clingo
@@ -104,25 +104,6 @@ class FieldTestCase(unittest.TestCase):
         self.assertEqual(RawField.pytocl(Raw(symstr)), symstr)
         self.assertEqual(RawField.pytocl(symstr), symstr)
 
-        symstr = Number(0)
-        self.assertEqual(type(BooleanField.cltopy(symstr)), bool)
-        self.assertEqual(BooleanField.cltopy(symstr), False)
-        self.assertEqual(BooleanField.pytocl(0), symstr)
-
-        symstr = String("false")
-        self.assertEqual(type(BooleanField.cltopy(symstr)), bool)
-        self.assertEqual(BooleanField.cltopy(symstr), False)
-
-        self.assertEqual(BooleanField.pytocl("false"), Number(0))
-        self.assertEqual(BooleanField.pytocl("TrUe"), Number(1))
-        self.assertEqual(BooleanField.pytocl("on"), Number(1))
-        self.assertEqual(BooleanField.pytocl("off"), Number(0))
-
-        symstr = Number(1)
-        self.assertEqual(type(StrictBooleanField.cltopy(symstr)), bool)
-        self.assertEqual(StrictBooleanField.cltopy(symstr), True)
-        self.assertEqual(StrictBooleanField.pytocl(True), symstr)
-
         # Now some bad conversions
         with self.assertRaises(TypeError) as ctx:
             x=StringField.cltopy(Number(1))
@@ -135,14 +116,6 @@ class FieldTestCase(unittest.TestCase):
         with self.assertRaises(TypeError) as ctx:
             x=ConstantField.cltopy(Function("x",[Number(1)]))
         check_errmsg("Symbol 'x(1)'",ctx)
-
-        with self.assertRaises(TypeError) as ctx:
-            x=BooleanField.cltopy(String("2"))
-        check_errmsg("value '2'",ctx)
-
-        with self.assertRaises(TypeError) as ctx:
-            x=StrictBooleanField.cltopy(Number(2))
-        check_errmsg("value must be either",ctx)
 
     #--------------------------------------------------------------------------
     # Test that the simple field unify functions work as expected for clingo
